Oss release 28.0.0 created 2023-05-25-11-29
see CHANGELOG.md for details

Original commit sha: c95faa38a371a2f4cacfe5a7866700d134c64aac

+++ client/ramses-client/impl/RamsesObjectRegistryIterator.h
@@ -19,7 +19,7 @@ namespace ramses
     {
     public:
         RamsesObjectRegistryIterator(const RamsesObjectRegistry& registry, ERamsesObjectType type)
-            : m_objects(registry.m_objects[type])
+            : m_objects(registry.m_objects[static_cast<int>(type)])
             , m_objectsTotalCount(m_objects.getTotalCount())
             , m_current(0u)
         {
